third_party/WebKit/LayoutTests/crypto/subtle/hmac/export-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test exporting an AES key.");

jsTestIsAsync = true;

var extractable = true;
var nonExtractable = false;

var jwkKey = {
    "kty": "oct",
    "k": "ahjkn-_387fgnsibf23qsvahjkn-_387fgnsibf23qs"
};

debug("Importing a JWK key...");
crypto.subtle.importKey("jwk", jwkKey, { name: "HMAC", hash: {name: "SHA-256"} }, extractable, ["sign", "verify"]).then(function(result) {
    key = result;

    return crypto.subtle.exportKey(null, key);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
    return crypto.subtle.exportKey(undefined, key);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
    return crypto.subtle.exportKey({}, key);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
    return crypto.subtle.exportKey("", key);
}).then(failAndFinishJSTest, function(result) {
    logError(result);
    return crypto.subtle.exportKey("foobar", key);
}).then(failAndFinishJSTest, function(result) {
    logError(result);

    debug("Exporting the key as raw data...");
    return crypto.subtle.exportKey("raw", key);
}, failAndFinishJSTest).then(function(result) {
    exportedData = result;
    shouldBe("bytesToHexString(new Uint8Array(exportedData))", "'6a18e49feff7f3b7e09ec89b7f6deab2f6a18e49feff7f3b7e09ec89b7f6deab'");

    debug("Exporting the key as JWK...");
    return crypto.subtle.exportKey("jwk", key);
}).then(function(result) {
    exportedJWK = result;
    shouldBe("exportedJWK.kty", "'oct'");
    shouldBe("exportedJWK.k", "'ahjkn-_387fgnsibf23qsvahjkn-_387fgnsibf23qs'");
    shouldBe("exportedJWK.alg", "'HS256'");
    shouldBe("exportedJWK.ext", "true");
    shouldBe("exportedJWK.use", "undefined");
    shouldBe("exportedJWK.key_ops", "['sign', 'verify']");

    debug("\nImporting a key that's not extractable...");
    return crypto.subtle.importKey("jwk", jwkKey, { name: "HMAC", hash: {name: "SHA-256"} }, nonExtractable, ["sign", "verify"]);
}, failAndFinishJSTest).then(function(result) {
    key = result;

    debug("\nTrying to export as raw...");
    return crypto.subtle.exportKey("raw", key);
}).then(function(result) {
    testFailed("Promise wasn't rejected");
    finishJSTest();
}, function(result) {
    logError(result);
    testPassed("Rejected, as expected");

    debug("Trying to export as jwk...");
    return crypto.subtle.exportKey("jwk", key);
}).then(function(result) {
    testFailed("Promise wasn't rejected");
    finishJSTest();
}, function(result) {
    logError(result);
    testPassed("Rejected, as expected");

    finishJSTest();
});
</script>

</body>
</html>
